# agricultural/api-tests/apifox-openapi-import.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "助农电商平台 API",
    "description": "用户端 / 商家端 / 管理员端 RESTful 接口 - 助农电商平台",
    "contact": {
      "name": "助农平台技术支持",
      "email": "support@agricultural.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/merchant/sms/send": {
      "post": {
        "tags": ["商家端-登录注册"],
        "summary": "发送短信验证码",
        "operationId": "sendMerchantSms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {"type": "string"}
                },
                "required": ["phone"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/merchant/register": {
      "post": {
        "tags": ["商家端-登录注册"],
        "summary": "商家注册",
        "operationId": "merchantRegister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {"type": "string"},
                  "code": {"type": "string"},
                  "merchantName": {"type": "string"},
                  "contactPerson": {"type": "string"},
                  "contactPhone": {"type": "string"},
                  "password": {"type": "string"}
                },
                "required": ["phone", "code", "merchantName", "contactPerson", "contactPhone"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultMerchantLoginVO"}}}}
        }
      }
    },
    "/merchant/login": {
      "post": {
        "tags": ["商家端-登录注册"],
        "summary": "商家登录",
        "operationId": "merchantLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {"type": "string"},
                  "password": {"type": "string"}
                },
                "required": ["phone", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultMerchantLoginVO"}}}}
        }
      }
    },
...
    "/merchant/orders": {
      "get": {
        "tags": ["商家端-订单管理"],
        "summary": "订单列表",
        "operationId": "merchantOrders",
        "parameters": [
          {"name": "orderStatus", "in": "query", "schema": {"type": "integer"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultMerchantOrderVO"}}}}
        }
      }
    },
    "/merchant/orders/{id}": {
      "get": {
        "tags": ["商家端-订单管理"],
        "summary": "订单详情",
        "operationId": "merchantOrderDetail",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultMerchantOrderVO"}}}}
        }
      }
    },
    "/merchant/orders/{id}/ship": {
      "put": {
        "tags": ["商家端-订单管理"],
        "summary": "发货",
        "operationId": "shipOrder",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "logisticsCompany": {"type": "string"},
                  "logisticsNo": {"type": "string"}
                },
                "required": ["logisticsCompany", "logisticsNo"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/merchant/orders/{id}/cancel": {
      "put": {
        "tags": ["商家端-订单管理"],
        "summary": "取消订单",
        "operationId": "merchantCancelOrder",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/merchant/after-sale/{no}/messages": {
      "get": {
        "tags": ["商家端-售后沟通"],
        "summary": "沟通列表",
        "operationId": "merchantAfterSaleMessages",
        "parameters": [
          {"name": "no", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/AfterSaleCommunication"}}}}}
        }
      },
      "post": {
        "tags": ["商家端-售后沟通"],
        "summary": "发送消息",
        "operationId": "sendMerchantAfterSaleMessage",
        "parameters": [
          {"name": "no", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {"type": "string"}
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": ["管理员端-用户管理"],
        "summary": "用户列表",
        "operationId": "adminUserList",
        "parameters": [
          {"name": "status", "in": "query", "schema": {"type": "integer"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultUserInfo"}}}}
        }
      },
      "post": {
        "tags": ["管理员端-用户管理"],
        "summary": "新增用户",
        "operationId": "addAdminUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {"type": "string"},
                  "nickname": {"type": "string"},
                  "password": {"type": "string"},
                  "status": {"type": "integer"}
                },
                "required": ["phone"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultUserInfo"}}}}
        }
      }
    },
    "/admin/users/{id}/status": {
      "put": {
        "tags": ["管理员端-用户管理"],
        "summary": "更新用户状态",
        "operationId": "updateUserStatus",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {"type": "integer", "description": "0-禁用 1-正常"}
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/admin/users/{id}": {
      "delete": {
        "tags": ["管理员端-用户管理"],
        "summary": "删除用户",
        "operationId": "deleteUser",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/admin/payments": {
      "get": {
        "tags": ["管理员端-支付管理"],
        "summary": "支付记录列表",
        "operationId": "paymentList",
        "parameters": [
          {"name": "payStatus", "in": "query", "schema": {"type": "integer"}},
          {"name": "refundStatus", "in": "query", "schema": {"type": "integer"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultPaymentRecord"}}}}
        }
      }
    },
    "/admin/payments/{id}/refund": {
      "put": {
        "tags": ["管理员端-支付管理"],
        "summary": "更新退款状态",
        "operationId": "updateRefundStatus",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refundStatus": {"type": "integer", "description": "0-未退款 1-已退款"}
                },
                "required": ["refundStatus"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"type": "object"}
        }
      },
...
      "ResultUserInfo": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/UserInfo"}
        }
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "phone": {"type": "string"},
          "nickname": {"type": "string"},
          "avatarUrl": {"type": "string"},
          "status": {"type": "integer"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
...
      "AfterSaleCommunication": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "afterSaleNo": {"type": "string"},
          "senderType": {"type": "integer", "description": "1-用户 2-商家 3-管理员"},
          "senderId": {"type": "integer", "format": "int64"},
          "content": {"type": "string"},
          "isRead": {"type": "integer"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
...
      "ResultMerchantLoginVO": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/MerchantLoginVO"}
        }
      },
      "MerchantLoginVO": {
        "type": "object",
        "properties": {
          "token": {"type": "string"},
          "merchantId": {"type": "integer", "format": "int64"},
          "merchantName": {"type": "string"},
          "auditStatus": {"type": "integer"},
          "status": {"type": "integer"}
        }
      },
...
      "MerchantOrderVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "orderNo": {"type": "string"},
          "userId": {"type": "integer", "format": "int64"},
          "totalAmount": {"type": "number"},
          "orderStatus": {"type": "integer"},
          "receiver": {"type": "string"},
          "receiverPhone": {"type": "string"},
          "receiverAddress": {"type": "string"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "PageResultMerchantOrderVO": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/MerchantOrderVO"}}
        }
      },
      "ResultMerchantOrderVO": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/MerchantOrderVO"}
        }
      },
...
      "PaymentRecord": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "orderNo": {"type": "string"},
          "payType": {"type": "integer"},
          "payAmount": {"type": "number"},
          "payStatus": {"type": "integer"},
          "payTime": {"type": "string", "format": "date-time"},
          "refundStatus": {"type": "integer"},
          "refundAmount": {"type": "number"}
        }
      },
      "PageResultPaymentRecord": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/PaymentRecord"}}
        }
      },
...
    }
  }
}