# agricultural/src/main/java/com/agricultural/assistplatform/controller/admin/AdminUserController.java
package com.agricultural.assistplatform.controller.admin;

import com.agricultural.assistplatform.common.PageResult;
import com.agricultural.assistplatform.common.Result;
import com.agricultural.assistplatform.entity.UserInfo;
import com.agricultural.assistplatform.service.admin.AdminUserService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;

import java.util.Map;

@Tag(name = "管理员端-用户管理")
@RestController
@RequestMapping("/admin")
@RequiredArgsConstructor
public class AdminUserController {

    private final AdminUserService adminUserService;

    @Operation(summary = "用户列表")
    @GetMapping("/users")
    public Result<PageResult<UserInfo>> list(
            @RequestParam(required = false) Integer status,
            @RequestParam(defaultValue = "1") Integer pageNum,
            @RequestParam(defaultValue = "10") Integer pageSize) {
        return Result.ok(adminUserService.list(status, pageNum, pageSize));
    }

    @Operation(summary = "新增用户")
    @PostMapping("/users")
    public Result<Map<String, Object>> create(@RequestBody UserInfo user) {
        Long id = adminUserService.create(user);
        return Result.ok(Map.of("id", id));
    }

    @Operation(summary = "更新用户状态")
    @PutMapping("/users/{id}/status")
    public Result<Void> updateStatus(@PathVariable Long id, @RequestBody Map<String, Integer> body) {
        Integer status = body != null ? body.get("status") : null;
        adminUserService.updateStatus(id, status);
        return Result.ok();
    }

    @Operation(summary = "删除用户")
    @DeleteMapping("/users/{id}")
    public Result<Void> delete(@PathVariable Long id) {
        adminUserService.delete(id);
        return Result.ok();
    }
}